+++ src/Data/Profunctor/Grammar.hs
@@ -40,7 +40,6 @@ import Data.Profunctor.Monoidal
 import Data.Void
 import Prelude hiding (id, (.))
 import GHC.Exts
-import Witherable
 
 -- | `Printor` is a simple printer `Profunctor`.
 newtype Printor s f a b = Printor {runPrintor :: a -> f (b, s -> s)}
@@ -180,7 +179,8 @@ instance
 instance BackusNaurForm (Parsor s m a b)
 instance (Alternative m, Monad m) => MonadFail (Parsor s m a) where
   fail _ = empty
-instance (Alternative m, Monad m) => MonadTry (Parsor s m a)
+instance (Alternative m, Monad m, Filterable m)
+  => MonadTry (Parsor s m a)
 instance AsEmpty s => Matching s (Parsor s [] a b) where
   word =~ p = case
     [ () | (_, remaining) <- runParsor p Nothing word
@@ -289,7 +289,8 @@ instance
 instance BackusNaurForm (Printor s m a b)
 instance (Alternative m, Monad m) => MonadFail (Printor s m a) where
   fail _ = empty
-instance (Alternative m, Monad m) => MonadTry (Printor s m a)
+instance (Alternative m, Monad m, Filterable m)
+  => MonadTry (Printor s m a)
 
 -- Grammor instances
 instance Functor (Grammor k a) where fmap _ = coerce